# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API для заметок с JWT-аутентификацией, refresh-сессиями и OAuth через Yandex.\nОсновной runtime использует HttpOnly cookies, а схема BearerAuth добавлена в Swagger UI для ручного тестирования защищенных методов.",
        "title": "Web Labs API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@weblab.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:4200",
    "basePath": "/",
    "paths": {
...
        "/notes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список заметок текущего пользователя с пагинацией",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Список заметок",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Размер страницы",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список заметок",
                        "schema": {
                            "$ref": "#/definitions/service.NotesResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает новую заметку, принадлежащую авторизованному пользователю",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Создать заметку",
                "parameters": [
                    {
                        "description": "Данные заметки",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreateNoteDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Заметка создана",
                        "schema": {
                            "$ref": "#/definitions/service.NoteResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные входные данные",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/notes/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает заметку по идентификатору, если она принадлежит текущему пользователю",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Получить заметку",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID заметки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заметка найдена",
                        "schema": {
                            "$ref": "#/definitions/service.NoteResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Доступ к чужой заметке запрещен",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Заметка не найдена",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Полностью заменяет данные заметки текущего пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Полностью обновить заметку",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID заметки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные заметки",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UpdateNoteDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заметка обновлена",
                        "schema": {
                            "$ref": "#/definitions/service.NoteResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные входные данные",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Доступ к чужой заметке запрещен",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Заметка не найдена",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Помечает заметку как удаленную, не удаляя запись физически из базы",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Удалить заметку",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID заметки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Заметка удалена"
                    },
                    "400": {
                        "description": "Некорректный ID",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Доступ к чужой заметке запрещен",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Заметка не найдена",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Частично обновляет поля заметки текущего пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Частично обновить заметку",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID заметки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Изменяемые поля заметки",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UpdateNoteDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заметка обновлена",
                        "schema": {
                            "$ref": "#/definitions/service.NoteResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные входные данные",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Доступ к чужой заметке запрещен",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Заметка не найдена",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "httpapi.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        },
...
        "service.CreateNoteDTO": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Текст заметки"
                },
                "title": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1,
                    "example": "Первая заметка"
                }
            }
        },
        "service.NoteResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Текст заметки"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-05-02T12:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "title": {
                    "type": "string",
                    "example": "Первая заметка"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-05-02T12:10:00Z"
                }
            }
        },
        "service.NotesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.NoteResponse"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/service.PaginationMeta"
                }
            }
        },
        "service.PaginationMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 10
                },
                "total_pages": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "service.UpdateNoteDTO": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Обновленный текст"
                },
                "title": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1,
                    "example": "Обновленная заметка"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите access token в формате \"Bearer {token}\". В приложении токен обычно передается через HttpOnly cookie access_token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}